examples/zillow-api-challenge/requirements.txt
Django==2.1.5
django-filter==2.1.0
djangorestframework==3.9.1
pytz==2018.9
